[features] добавил возможность записи на несколько часов подряд, также теперь нельзя записываться на прошедшие даты и отменять записи с прошедших дат

+++ src/main/java/ru/kukulo1/test_assignment/reservation/ReservationService.java
@@ -12,6 +12,7 @@ import ru.kukulo1.test_assignment.sessionhour.SessionHourRepository;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -37,13 +38,17 @@ public class ReservationService {
             return new ResponseEntity<>("Клиента с таким ID не найдено!", HttpStatus.BAD_REQUEST);
         }
 
-        if (!reservationRepository.findByClientIdAndDate(addReservationRecord.clientID(), addReservationRecord.localDateTime().toLocalDate()).isEmpty()) {
+        if (addReservationRecord.dateTime().toLocalDate().isBefore(LocalDate.now())) {
+            return new ResponseEntity<>("Невозможно записаться на прошедшую дату!", HttpStatus.BAD_REQUEST);
+        }
+
+        if (!reservationRepository.findByClientIdAndDate(addReservationRecord.clientID(), addReservationRecord.dateTime().toLocalDate()).isEmpty()) {
             return new ResponseEntity<>(String.format("Попытка второй записи клиента с ID: %s за день!", addReservationRecord.clientID()), HttpStatus.BAD_REQUEST);
         }
 
 
-        if (reservationRepository.isSlotAvailable(addReservationRecord.localDateTime())) {
-            Optional<SessionHour> sessionHourOpt = sessionHourRepository.findByDateTime(addReservationRecord.localDateTime());
+        if (reservationRepository.isSlotAvailable(addReservationRecord.dateTime())) {
+            Optional<SessionHour> sessionHourOpt = sessionHourRepository.findByDateTime(addReservationRecord.dateTime());
 
             if (sessionHourOpt.isPresent()) {
                 reservationRepository.save(new Reservation(
@@ -51,11 +56,11 @@ public class ReservationService {
                         sessionHourOpt.get()
                 ));
                 return new ResponseEntity<>(String.format("Клиент с ID: %s записан на %s",
-                        addReservationRecord.clientID(), addReservationRecord.localDateTime()), HttpStatus.OK);
+                        addReservationRecord.clientID(), addReservationRecord.dateTime()), HttpStatus.OK);
             }
         }
 
-        if (doesTimestampHasMinutes(addReservationRecord.localDateTime())) {
+        if (doesTimestampHasMinutes(addReservationRecord.dateTime())) {
             return new ResponseEntity<>("Попытка записаться на неполный час!", HttpStatus.BAD_REQUEST);
         } else {
             return new ResponseEntity<>("Попытка записаться на нерабочее время!", HttpStatus.BAD_REQUEST);
@@ -74,6 +79,10 @@ public class ReservationService {
         }
         if (!reservation.get().getClient().getId().equals(cancelReservationRecord.clientID())) {
             return new ResponseEntity<>(String.format("Запись с ID: %s не принадлежит клиенту с ID: %s. Отмена невозможна!", cancelReservationRecord.reservationID(), cancelReservationRecord.clientID()), HttpStatus.BAD_REQUEST);
+        }
+
+        if (reservation.get().getSessionHour().getDateTime().toLocalDate().isBefore(LocalDate.now())) {
+            return new ResponseEntity<>("Невозможно отменить запись с прошедшей даты!", HttpStatus.BAD_REQUEST);
         }
 
         reservationRepository.delete(reservation.get());
@@ -107,7 +116,62 @@ public class ReservationService {
         return new ResponseEntity<>(reservationsRecords, HttpStatus.OK);
     }
 
+    public ResponseEntity<String> reserveSessionHourInterval(AddReservationForSeveralHoursRecord addReservationForSeveralHoursRecord) {
+        if (clientRepository.findById(addReservationForSeveralHoursRecord.clientID()).isEmpty()) {
+            return new ResponseEntity<>("Клиента с таким ID не найдено!", HttpStatus.BAD_REQUEST);
+        }
+
+        if (addReservationForSeveralHoursRecord.dateTimeStart().toLocalDate().isBefore(LocalDate.now())) {
+            return new ResponseEntity<>("Невозможно записаться на прошедшую дату!", HttpStatus.BAD_REQUEST);
+        }
+
+        if (!reservationRepository.findByClientIdAndDate(addReservationForSeveralHoursRecord.clientID(), addReservationForSeveralHoursRecord.dateTimeStart().toLocalDate()).isEmpty()) {
+            return new ResponseEntity<>(String.format("Попытка второй записи клиента с ID: %s за день!", addReservationForSeveralHoursRecord.clientID()), HttpStatus.BAD_REQUEST);
+        }
+
+        if (addReservationForSeveralHoursRecord.dateTimeStart().isAfter(addReservationForSeveralHoursRecord.dateTimeEnd())) {
+            return new ResponseEntity<>("Время окончания записи раньше чем время её начала!", HttpStatus.BAD_REQUEST);
+        }
+
+        if (!addReservationForSeveralHoursRecord.dateTimeStart().toLocalDate().equals(addReservationForSeveralHoursRecord.dateTimeEnd().toLocalDate())) {
+            return new ResponseEntity<>("Дата начала записи и дата ёё окончания не совпадают!", HttpStatus.BAD_REQUEST);
+        }
+
+        List<LocalDateTime> sessionHours = getSlots(addReservationForSeveralHoursRecord.dateTimeStart(), addReservationForSeveralHoursRecord.dateTimeEnd());
+
+        for (LocalDateTime sessionHour : sessionHours) {
+            if (!reservationRepository.isSlotAvailable(sessionHour)) {
+                if (doesTimestampHasMinutes(addReservationForSeveralHoursRecord.dateTimeStart())) {
+                    return new ResponseEntity<>("Попытка записаться на неполный час!", HttpStatus.BAD_REQUEST);
+                } else {
+                    return new ResponseEntity<>("Попытка записаться на нерабочее время!", HttpStatus.BAD_REQUEST);
+                }
+            }
+        }
+
+        for (LocalDateTime sessionHour : sessionHours) {
+            Optional<SessionHour> sessionHourOpt = sessionHourRepository.findByDateTime(sessionHour);
+            reservationRepository.save(new Reservation(
+                    clientRepository.findById(addReservationForSeveralHoursRecord.clientID()).get(),
+                    sessionHourOpt.get()
+            ));
+        }
+        return new ResponseEntity<>(String.format("Клиент с ID: %s записан на %s - %s",
+                addReservationForSeveralHoursRecord.clientID(), addReservationForSeveralHoursRecord.dateTimeStart(), addReservationForSeveralHoursRecord.dateTimeEnd()), HttpStatus.OK);
+    }
+
     private boolean doesTimestampHasMinutes(LocalDateTime localDateTime) {
         return localDateTime.toLocalTime().getMinute() != 0;
+    }
+    private List<LocalDateTime> getSlots(LocalDateTime dateTimeStart, LocalDateTime dateTimeEnd) {
+        List<LocalDateTime> slots = new ArrayList<>();
+        LocalDateTime current = dateTimeStart;
+
+        while (current.isBefore(dateTimeEnd)) {
+            slots.add(current);
+            current = current.plusHours(1);
+        }
+
+        return slots;
     }
 }
